requirements/base.txt
# postgres engine
psycopg2-binary==2.8.4

# Django
Django==2.2.10 # https://www.djangoproject.com/

# enable cross-site resource sharing with frontend
django-cors-headers==2.4.0 # https://github.com/ottoyiu/django-cors-headers/

# enable dynamic filter
django-filter==2.2.0 # https://github.com/carltongibson/django-filter

# all sorts of django extras like shell_plus
django-extensions==2.2.8 # https://github.com/django-extensions/django-extensions

# django jwt token authentication
djangorestframework-simplejwt==4.4.0

# emailer
django-smtp-ssl==1.0 # https://github.com/bancek/django-smtp-ssl

# ipython
ipython==7.12.0

# django rest framework
djangorestframework==3.11.0
django-rest-auth==0.9.5
